CMSIS-NN: Partial fix for failing conv unit test cases
Fixes #43475 when combined with PR that fixes #42951

Fix #42951 ensures that the reference implementation is invoked
when dilation is not equal to 1 as that case does not have an
optimized implementation.

This PR addresses the issue that zero point arguments are
uninitialized when invoking the reference implementation of
convolution resulting in failing test cases.