Fix the build for benchmarks for bluepill and STM32F4.
The root cause here is that a static inside a function still has locks
around it, but a global static does noti. This is the behavior if we use
no-threadsafe-statics.

Also note that while this change fixes the build, the currentHEAD is
broken because of additiona changes in the tree. See issue #43510 for
more details.

We are going to improve the coverage of the CI system with PR #43509 and
are going to merge this change separately to better root-cause the build
breakage and motivate the reverts that will be necessary.

Manually tested with the following commands:

make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=stm32f4 TAGS=cmsis-nn keyword_benchmark person_detection_benchmark persone_detection_experimental_benchmark
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=bluepill keyword_benchmark person_detection_benchmark persone_detection_experimental_benchmark