Add plumbing for setting enable_mlir_bridge=False
In order to roll out the mlir_bridge, we need to support forcibly
enabling the bridge, forcibly disabling the bridge and a default state.
If the bridge is not forciby enabled or disabled, the new bridge will
decide whether or not it should be enabled. Overtime, the new bridge
will increasing select more models to run.

This cl adds front-end support for forcibly disabling the bridge.
For now, if the bride is not forcibly enabled, it is assumed to
be disabled.

PiperOrigin-RevId: 333767292
Change-Id: Icbc89dd6e7a49c98963d157bd030cda8e5b262ed