[TF:TRT] Fix TF-TRT python integration tests.
Routine _RunGraph needs a session config to support the running of an inference
graph with TF1 for the convert_online mode. Previously, we always pass a session
config to _RunGraph. This has two problems. One is when we run the original
graph to get the reference result with TF1, we actually have TF-TRT enabled and
run the TF-TRT converted graph by mistake. Another is when we run the TF-TRT
converted inference graph, we run the TF-TRT optimization again on the
converted graph. This change delays the creation of a session config until it is
needed inside _RunGraph and uses the tf.session default config None for all
other cases to fix the problem.

PiperOrigin-RevId: 335036016
Change-Id: Icd8ebc5db999dbd41da3d9c45f17d3cac1cfd398