PR #43509: Build benchmarks as part of `make build`
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/43509

This should make it easier to ensure that benchmarks can be run on a variety of platforms.

Note that in light of #43510 we may additionally want to build bluepill with TAGS=cmsis-nn but keeping that as a separate change in the interest of incrementally improving our CI coverage.

Manually confirmed that `make build` now also builds the benchmarks.

Copybara import of the project:

--
c0052e8f095ebfe8996a192e39898e744cbd70bc by Advait Jain <advaitjain@users.noreply.github.com>:

Build benchmarks as part of `make build`

This should make it easier to ensure that benchmarks can be run on a
variety of platforms.

PiperOrigin-RevId: 335328522
Change-Id: I5420440e666ffa815729d5167ddc0ef9d3637474