Save FunctionSpec with BareConcreteFunction
ConcreteFunction now supports being called with both structured inputs and flatten inputs, so that it's compatible with tf.function calling semantics.

To support the same semantics for saved ConcreteFunction, we need to save the FunctionSpec.

Note that calling with structured inputs is not yet supported in SavedModel C++ API.

This change also modifies ConcreteFunction.pretty_printed_signature to make it work for loaded functions. The structured_outputs of loaded functions are tensor specs instead of tensors. Ideally it should be tensors as well, but there're already users who depend on
this behavior.

PiperOrigin-RevId: 335946423
Change-Id: I4aecf2aba51459801bd0d42a3343ad1becaef187