[TF:TRT] Ensure that each TensorRT layer has a non-empty name to support xprof.
Extend the layer name format from [main_op_name]_[sub_op_name] to
[main_op_name]_[sub_op_name]_[sub_op_instance]. This is to handle the case
where we need to generate a few layers with the same operation type to convert
a TF operation to TensorRT, and the number of such layers is only known at
conversion time. An example of this case is a TF node may be transposed in a
few ways to produce a few inputs of a TRTEngineOp. During conversion, we will
generate a few shuffle layers for the TF node and the name of those shuffle
layers only differ in their [sub_op_instance]. This avoids having different
layers to share the same name, which causes TensorRT engine creation failure.

Extend a few utility routines that may be called a few times to create TensorRT
layers when converting the same TF node with a parameter to represent the call
instance. This allows us to use the instance index as part of the TensorRT
layer names to avoid name collision.

Fix the operation converters to set up TensorRT layer names.

Modify convert_nodes_test to check layer names are not empty.

PiperOrigin-RevId: 336697580
Change-Id: Ie1f1c80fd9f8da9532e722d84f1bbbeda210ec73