[tf.data service] Fix issue where finished tasks are restarted.
When a worker finishes a task, it asynchronously informs the dispatcher. The worker also sends periodic heartbeats to keep its tasks in sync with what the dispatcher thinks its tasks should be. This led to the following race condition:

1. Worker finishes task T.
2. Worker erases T.
3. Worker heartbeats to the dispatcher. Dispatcher sees that the worker is not working on T, and tells it to start working on T. Worker starts T from the beginning.
4. Consumers read additional elements from T, when they should be getting end_of_sequence instead.

To fix this, the worker will now mark tasks as finished when they finish, and wait to erase them until the dispatcher tells the worker to erase them.

This issue was causing some unit tests to fail flakily due to the race condition.

PiperOrigin-RevId: 337197366
Change-Id: Ie9053af37c8f6200a45441bcb8e84e30daf40057