[tf.data service] Retry stream removed errors.
When auto-reconnecting to a tf.data server after it restarts, gRPC can return an UNKNOWN error code with a "Stream removed" error message. This should not be treated as an unrecoverable error. The `FromGrpcStatus` util method has logic to convert these UNKNOWN "Stream removed" errors into UNAVAILABLE errors so that they will be retried.

PiperOrigin-RevId: 337589493
Change-Id: I2cbf9071d9b8a4ada34e97982d0c5fecfaaf0220