Have AutoCastVariable.dtype refer to the variable dtype.
This allows us the flexibility to later remove AutoCastVariable and instead have a mechanism so that individual ops will cast variables (and potentially other tensors) to the correct dtype. See the last paragraph of the this section of the mixed precision RFC (https://github.com/tensorflow/community/blob/85635744559cdfe27ff92bf11adbd17c6f88ae73/rfcs/20200929-keras-mixed-precision.md#op-based-autocasting-api) for an example of how this could be done.

PiperOrigin-RevId: 337793570
Change-Id: I8e56f7d276117a9a81070ab0984369e8a4490eea